requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.2.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.115.6
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==16.0
